refactor(sidebar): extract chat management logic into useSidebarChats hook for improved code organization and reusability
