+++ components/sidebar/use-sidebar-chats.ts
@@ -1,0 +1,82 @@
+import { useCallback, useEffect, useState } from 'react';
+
+type Chat = {
+  id: string;
+  title: string;
+};
+
+export default function useSidebarChats(initialChats: Chat[]) {
+  const [chats, setChats] = useState<Chat[]>(initialChats);
+
+  const refreshChats = useCallback(async () => {
+    const res = await fetch('/api/chats', {
+      cache: 'no-store',
+    });
+
+    if (!res.ok) {
+      console.error('Load chats failed');
+      return;
+    }
+
+    const data: Chat[] = await res.json();
+    setChats(data);
+  }, []);
+
+  useEffect(() => {
+    const handleRefresh = () => {
+      void refreshChats();
+    };
+
+    window.addEventListener('chats:refresh', handleRefresh);
+
+    return () => {
+      window.removeEventListener('chats:refresh', handleRefresh);
+    };
+  }, [refreshChats]);
+
+  const renameChat = async (chatId: string, title: string) => {
+    const res = await fetch(`/api/chats/${chatId}`, {
+      method: 'PATCH',
+      headers: {
+        'Content-Type': 'application/json',
+      },
+      body: JSON.stringify({ title }),
+    });
+
+    if (!res.ok) {
+      console.error('Rename chat failed');
+      return;
+    }
+
+    const updatedChat = await res.json();
+
+    setChats((prev) =>
+      prev.map((chat) => (chat.id === chatId ? { ...chat, title: updatedChat.title } : chat)),
+    );
+  };
+
+  const deleteChat = async (chatId: string) => {
+    const confirmed = window.confirm('Delete this chat?');
+
+    if (!confirmed) return false;
+
+    const res = await fetch(`/api/chats/${chatId}`, {
+      method: 'DELETE',
+    });
+
+    if (!res.ok) {
+      console.error('Delete chat failed');
+      return false;
+    }
+
+    setChats((prev) => prev.filter((chat) => chat.id !== chatId));
+    return true;
+  };
+
+  return {
+    chats,
+    deleteChat,
+    refreshChats,
+    renameChat,
+  };
+}
